feat: Improve post page display
This commit improves the display of the post page by making the author and date information conditional. If the post has an author, it will be displayed along with the post date. Otherwise, only the post date will be shown. This change enhances the readability and clarity of the post page. Also, the Block checking section is added to. So, the scrolling animation will be completed (initial)