feat: Update font import in og/route.tsx and FooterMain component
This commit updates the font import in the og/route.tsx file to use the GeistMono-Bold.ttf font file instead of ZedMonoNerdFontPropo-Extended.ttf. This change improves the visual consistency of the application.

Additionally, the commit updates the EULA link in the FooterMain component to use the Next.js Link component. This improves the navigation experience by leveraging client-side routing instead of full page reloads.